Order-Orderline ( 6/26/2026 )

+++ src/main/java/co/istad/sengkim/ite3rdecommerce/features/product/Product.java
@@ -28,7 +28,7 @@ public class Product {
     @Column(nullable = false)
     private String thumbnail;
     @Column(nullable = false)
-    private Integer unitPrice;
+    private Double unitPrice;
     @Column(nullable = false)
     private Integer qty;
